minor: rename folders
