+++ client/src/components/welcomePage/utils/index.ts
