+++ client/src/components/welcomePage/utils/welcomeText.tsx
