+++ client/src/components/welcomePage/welcomePage.scss
