+++ client/src/components/welcomePage/welcomePage.test.jsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { ThemeProvider } from '../../contexts/theme';
 import { UserProvider } from '../../contexts/user';
-import Welcome from './welcome'
+import Welcome from './welcomePage'
 import userEvent from "@testing-library/user-event";
 
 const MockWrapper = ({ children }) => (
